+++ packages/rs-drive/src/common/mod.rs
@@ -50,11 +50,9 @@ use grovedb::TransactionArg;
 use crate::contract::Contract;
 use crate::drive::Drive;
 use crate::error::Error;
-use dpp::data_contract::errors::structure::StructureError;
-use dpp::data_contract::extra::common::json_document_to_cbor;
+use dpp::data_contract::extra::common::{bytes_for_system_value, json_document_to_cbor};
 
 use crate::drive::block_info::BlockInfo;
-use dpp::data_contract::extra::DriveContractExt;
 
 /// Serializes to CBOR and applies to Drive a JSON contract from the file system.
 pub fn setup_contract(
@@ -154,20 +152,6 @@ pub fn cbor_map_to_btree_map(cbor_map: &[(Value, Value)]) -> BTreeMap<String, &V
         .collect::<BTreeMap<String, &Value>>()
 }
 
-/// Converts the keys which are dynamic CBOR strings from a CBOR map to a BTree map.
-pub fn cbor_owned_map_to_btree_map(cbor_map: Vec<(Value, Value)>) -> BTreeMap<String, Value> {
-    cbor_map
-        .into_iter()
-        .filter_map(|(key, value)| {
-            if let Value::Text(key) = key {
-                Some((key, value))
-            } else {
-                None
-            }
-        })
-        .collect::<BTreeMap<String, Value>>()
-}
-
 /// Retrieves the value of a key from a CBOR map if it's an array of strings.
 pub fn cbor_inner_array_of_strings<'a>(
     document_type: &'a [(Value, Value)],
@@ -250,226 +234,6 @@ pub fn cbor_inner_text_value<'a>(
     let key_value = get_key_from_cbor_map(document_type, key)?;
     if let Value::Text(string_value) = key_value {
         return Some(string_value);
-    }
-    None
-}
-
-/// Retrieves the value of a key from a B-tree map if it's a string.
-pub fn btree_map_inner_text_value<'a>(
-    document_type: &'a BTreeMap<String, &'a Value>,
-    key: &'a str,
-) -> Option<&'a str> {
-    let key_value = document_type.get(key)?;
-    if let Value::Text(string_value) = key_value {
-        return Some(string_value);
-    }
-    None
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a byte array.
-pub fn cbor_inner_bytes_value<'a>(
-    document_type: &'a [(Value, Value)],
-    key: &'a str,
-) -> Option<Vec<u8>> {
-    let key_value = get_key_from_cbor_map(document_type, key)?;
-    match key_value {
-        Value::Bytes(bytes) => Some(bytes.clone()),
-        Value::Array(array) => {
-            match array
-                .iter()
-                .map(|byte| match byte {
-                    Value::Integer(int) => {
-                        let value_as_u8: u8 = (*int).try_into().map_err(|_| {
-                            Error::Structure(StructureError::ValueWrongType("expected u8 value"))
-                        })?;
-                        Ok(value_as_u8)
-                    }
-                    _ => Err(Error::Structure(StructureError::ValueWrongType(
-                        "not an array of integers",
-                    ))),
-                })
-                .collect::<Result<Vec<u8>, Error>>()
-            {
-                Ok(bytes) => Some(bytes),
-                Err(_) => None,
-            }
-        }
-        _ => None,
-    }
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a boolean.
-pub fn cbor_inner_bool_value(document_type: &[(Value, Value)], key: &str) -> Option<bool> {
-    let key_value = get_key_from_cbor_map(document_type, key)?;
-    if let Value::Bool(bool_value) = key_value {
-        return Some(*bool_value);
-    }
-    None
-}
-
-/// Retrieves the value of a key from a B-tree map if it's a boolean.
-pub fn btree_map_inner_bool_value(
-    document_type: &BTreeMap<String, &Value>,
-    key: &str,
-) -> Option<bool> {
-    let key_value = document_type.get(key)?;
-    if let Value::Bool(bool_value) = key_value {
-        return Some(*bool_value);
-    }
-    None
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a u8 value.
-pub fn cbor_inner_size_value(document_type: &[(Value, Value)], key: &str) -> Option<usize> {
-    let key_value = get_key_from_cbor_map(document_type, key)?;
-    if let Value::Integer(integer) = key_value {
-        let value_as_usize: Result<usize, Error> = (*integer)
-            .try_into()
-            .map_err(|_| Error::Structure(StructureError::ValueWrongType("expected u8 value")));
-        match value_as_usize {
-            Ok(size) => Some(size),
-            Err(_) => None,
-        }
-    } else {
-        None
-    }
-}
-
-/// Retrieves the value of a key from a B-tree map if it's a u8 value.
-pub fn btree_map_inner_size_value(
-    document_type: &BTreeMap<String, &Value>,
-    key: &str,
-) -> Option<usize> {
-    let key_value = document_type.get(key)?;
-    if let Value::Integer(integer) = key_value {
-        let value_as_usize: Result<usize, Error> = (*integer)
-            .try_into()
-            .map_err(|_| Error::Structure(StructureError::ValueWrongType("expected u8 value")));
-        match value_as_usize {
-            Ok(size) => Some(size),
-            Err(_) => None,
-        }
-    } else {
-        None
-    }
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a boolean otherwise returns a default boolean.
-pub fn cbor_inner_bool_value_with_default(
-    document_type: &[(Value, Value)],
-    key: &str,
-    default: bool,
-) -> bool {
-    cbor_inner_bool_value(document_type, key).unwrap_or(default)
-}
-
-/// Takes a value (should be a system value) and returns it as a byte array if possible.
-pub fn bytes_for_system_value(value: &Value) -> Result<Option<Vec<u8>>, Error> {
-    match value {
-        Value::Bytes(bytes) => Ok(Some(bytes.clone())),
-        Value::Text(text) => match bs58::decode(text).into_vec() {
-            Ok(data) => Ok(Some(data)),
-            Err(_) => Ok(None),
-        },
-        Value::Array(array) => array
-            .iter()
-            .map(|byte| match byte {
-                Value::Integer(int) => {
-                    let value_as_u8: u8 = (*int).try_into().map_err(|_| {
-                        Error::Structure(StructureError::ValueWrongType("expected u8 value"))
-                    })?;
-                    Ok(Some(value_as_u8))
-                }
-                _ => Err(Error::Structure(StructureError::ValueWrongType(
-                    "not an array of integers",
-                ))),
-            })
-            .collect::<Result<Option<Vec<u8>>, Error>>(),
-        _ => Err(Error::Structure(StructureError::ValueWrongType(
-            "system value is incorrect type",
-        ))),
-    }
-}
-
-/// Takes a B-tree map and a key and returns the corresponding value (should be a system value)
-/// as a byte array if possible.
-pub fn bytes_for_system_value_from_tree_map(
-    document: &BTreeMap<String, Value>,
-    key: &str,
-) -> Result<Option<Vec<u8>>, Error> {
-    let value = document.get(key);
-    if let Some(value) = value {
-        bytes_for_system_value(value)
-    } else {
-        Ok(None)
-    }
-}
-
-/// Takes a B-tree map, a key, and a default bool values and returns the corresponding
-/// value (should be a system value) from the key if it's a boolean, otherwise returns the default.
-pub fn bool_for_system_value_from_tree_map(
-    document: &BTreeMap<String, Value>,
-    key: &str,
-    default: bool,
-) -> Result<bool, Error> {
-    let value = document.get(key);
-    if let Some(value) = value {
-        if let Value::Bool(bool_value) = value {
-            Ok(*bool_value)
-        } else {
-            Err(Error::Structure(StructureError::ValueWrongType(
-                "value is expected to be a boolean",
-            )))
-        }
-    } else {
-        Ok(default)
-    }
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a u64.
-pub(crate) fn cbor_inner_u64_value<'a>(
-    document_type: &'a [(Value, Value)],
-    key: &'a str,
-) -> Option<u64> {
-    let key_value = get_key_from_cbor_map(document_type, key)?;
-    if let Value::Integer(integer_value) = key_value {
-        return Some(i128::from(*integer_value) as u64);
-    }
-    None
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a u32.
-pub(crate) fn cbor_inner_u32_value<'a>(
-    document_type: &'a [(Value, Value)],
-    key: &'a str,
-) -> Option<u32> {
-    let key_value = get_key_from_cbor_map(document_type, key)?;
-    if let Value::Integer(integer_value) = key_value {
-        return Some(i128::from(*integer_value) as u32);
-    }
-    None
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a u16.
-pub(crate) fn cbor_inner_u16_value<'a>(
-    document_type: &'a [(Value, Value)],
-    key: &'a str,
-) -> Option<u16> {
-    let key_value = get_key_from_cbor_map(document_type, key)?;
-    if let Value::Integer(integer_value) = key_value {
-        return Some(i128::from(*integer_value) as u16);
-    }
-    None
-}
-
-/// Retrieves the value of a key from a CBOR map if it's a u8.
-pub(crate) fn cbor_inner_u8_value<'a>(
-    document_type: &'a [(Value, Value)],
-    key: &'a str,
-) -> Option<u8> {
-    let key_value = get_key_from_cbor_map(document_type, key)?;
-    if let Value::Integer(integer_value) = key_value {
-        return Some(i128::from(*integer_value) as u8);
     }
     None
 }
