+++ packages/rs-drive/src/drive/query.rs
@@ -45,7 +45,6 @@ use dpp::data_contract::extra::DocumentType;
 
 use crate::drive::block_info::BlockInfo;
 use crate::fee_pools::epochs::Epoch;
-use dpp::data_contract::extra::DriveContractExt;
 
 impl Drive {
     /// Performs and returns the result of the specified query along with skipped items and the cost.
